NGLASS-525 Improve accordion performance
Accordion would previously try to render on any viewer change.
This was not very smart, because we would have to try render the
accordion each time.
There was also the case that, sometimes, we would see the elements
snap in to the accordion, sometimes requiring a tab switch for that
to happen.
Both issues were mitigated by moving the render of the accordion
into the tab selected event. This only triggers an accordion render
when we actually switch tabs.