fix: correct name of id selector for tool container
previously we were using the wrong name, so none of the styles were getting applied

Moved the styles under the neuroglass-theme selector as well, not sure why they were at the root level. Maybe the old selector used to spawn in the div at the wrong level, but if so, that has been fixed.